+++ src/services/UploadService.php
@@ -10,8 +10,10 @@
 namespace shardimage\shardimagephp\services;
 
 use shardimage\shardimagephpapi\api\Response;
+use shardimage\shardimagephp\builders\UploadBuilder;
 use shardimage\shardimagephp\helpers\ArrayHelper;
 use shardimage\shardimagephp\models\image\Image;
+use shardimage\shardimagephp\models\upload\UploadParams;
 use shardimage\shardimagephpapi\web\exceptions\NotImplementedHttpException;
 
 /**
@@ -22,7 +24,7 @@ class UploadService extends Service
     /**
      * Uploads an image from a local source.
      *
-     * @param array $params Required API parameters
+     * @param UploadParams|array $params Required API parameters
      *
      * <li>file - source (string - filepath, array - a $_FILES entry,
      * resource - an opened file resource (@see fopen()),
@@ -42,6 +44,10 @@ class UploadService extends Service
     public function upload($params, $optParams = [])
     {
         $resource = null;
+        if ($params instanceof UploadParams) {
+            $optParams = array_merge($optParams, $params->getOptionalParams());
+            $params = $params->getParams();
+        }
         if (!is_array($params)) {
             throw new \InvalidArgumentException('First parameter ($params) must to be an array!');
         }
@@ -80,7 +86,7 @@ class UploadService extends Service
     /**
      * Uploads an image from a remote source.
      *
-     * @param array $params Required API parameters
+     * @param UploadParams|array $params Required API parameters
      *
      * <li>resource - source (string - URL, an array with a 'remote' key consisting
      * of the above)
@@ -96,6 +102,10 @@ class UploadService extends Service
      */
     public function uploadRemote($params, $optParams = [])
     {
+        if ($params instanceof UploadParams) {
+            $optParams = array_merge($optParams, $params->getOptionalParams());
+            $params = $params->getParams();
+        }
         if (!is_array($params)) {
             throw new \InvalidArgumentException('First parameter ($params) must to be an array!');
         }
@@ -131,7 +141,7 @@ class UploadService extends Service
     /**
      * Modifies an image by changing storage format and/or transformations.
      *
-     * @param array|string $params Required API parameters
+     * @param UploadParams|array|string $params Required API parameters
      *
      * <li>publicId - image ID
      * <li>cloudId - cloud ID
@@ -146,6 +156,10 @@ class UploadService extends Service
      */
     public function modify($params, $optParams = [])
     {
+        if ($params instanceof UploadParams) {
+            $optParams = array_merge($optParams, $params->getOptionalParams());
+            $params = $params->getParams();
+        }
         if (is_string($params)) {
             $params = ['publicId' => $params];
         }