+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0-alpha37] - 2020-03-26
+### Add
+ - `nofw` URL option added to Option factory.
+ - `UploadBuilder` class added to support developer friendly image upload.
+
 ## [1.0.0-alpha36] - 2020-01-17
 ### Change
  - GET parameters at API request can be array.